Move db contents heuristics to separate file
