+++ extensions/ql-vscode/src/helpers.ts
@@ -5,7 +5,6 @@ import {
   writeFile,
   opendir,
 } from "fs-extra";
-import { glob } from "glob";
 import { join, basename, dirname } from "path";
 import { dirSync } from "tmp-promise";
 import { Uri, window as Window, workspace, env, WorkspaceFolder } from "vscode";
@@ -354,10 +353,6 @@ export async function prepareCodeTour(
 }
 
 /**
- * The following functions al heuristically determine metadata about databases.
- */
-
-/**
  * Returns the initial contents for an empty query, based on the language of the selected
  * databse.
  *
@@ -376,34 +371,6 @@ export function getInitialQueryContents(language: string, dbscheme: string) {
   }
 
   return language ? `import ${language}\n\nselect ""` : 'select ""';
-}
-
-/**
- * Heuristically determines if the directory passed in corresponds
- * to a database root. A database root is a directory that contains
- * a codeql-database.yml or (historically) a .dbinfo file. It also
- * contains a folder starting with `db-`.
- */
-export async function isLikelyDatabaseRoot(maybeRoot: string) {
-  const [a, b, c] = await Promise.all([
-    // databases can have either .dbinfo or codeql-database.yml.
-    pathExists(join(maybeRoot, ".dbinfo")),
-    pathExists(join(maybeRoot, "codeql-database.yml")),
-
-    // they *must* have a db-{language} folder
-    glob("db-*/", { cwd: maybeRoot }),
-  ]);
-
-  return (a || b) && c.length > 0;
-}
-
-/**
- * A language folder is any folder starting with `db-` that is itself not a database root.
- */
-export async function isLikelyDbLanguageFolder(dbPath: string) {
-  return (
-    basename(dbPath).startsWith("db-") && !(await isLikelyDatabaseRoot(dbPath))
-  );
 }
 
 export function isQueryLanguage(language: string): language is QueryLanguage {
